+++ projects/np-ui-lib/src/lib/np-checkbox/np-checkbox-group.component.ts
@@ -1,0 +1,13 @@
+import { Component, ViewEncapsulation, ChangeDetectionStrategy, Input } from '@angular/core';
+
+@Component({
+    selector: 'np-checkbox-group',
+    templateUrl: './np-checkbox-group.component.html',
+    styleUrls: ['./np-checkbox.component.css'],
+    encapsulation: ViewEncapsulation.None,
+    changeDetection: ChangeDetectionStrategy.Default
+})
+export class NpCheckboxGroupComponent {
+    // Orientation can be 'horizontal' or 'vertical'
+    @Input() orientation = 'horizontal';
+}