+++ projects/np-ui-lib/src/lib/np-checkbox/np-checkbox.component.ts
@@ -1,0 +1,101 @@
+import { ChangeDetectionStrategy, Component, ElementRef, EventEmitter, forwardRef, Input, Output, ViewChild, ViewEncapsulation } from '@angular/core';
+import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
+
+@Component({
+  selector: 'np-checkbox',
+  templateUrl: './np-checkbox.component.html',
+  styleUrls: ['./np-checkbox.component.css'],
+  encapsulation: ViewEncapsulation.None,
+  changeDetection: ChangeDetectionStrategy.Default,
+  providers: [
+    {
+      provide: NG_VALUE_ACCESSOR,
+      useExisting: forwardRef(() => NpCheckboxComponent),
+      multi: true
+    }
+  ]
+})
+export class NpCheckboxComponent implements ControlValueAccessor {
+  static controlCount = 1;
+
+  @Input() label: string;
+  @Input() indeterminate: boolean;
+  @Input() readOnly: boolean;
+  @Input() autoFocus: boolean;
+  @Input() tabIndex: number;
+  @Input() styleClass: string;
+  @Input() inputId = `np-checkbox_${NpCheckboxComponent.controlCount++}`;
+
+  @Output() onChange: EventEmitter<any> = new EventEmitter();
+  @Output() onFocus: EventEmitter<any> = new EventEmitter();
+  @Output() onBlur: EventEmitter<any> = new EventEmitter();
+
+  @ViewChild('control') inputViewChild: ElementRef;
+
+  innerValue: boolean;
+  isDisabled = false;
+  focused = false;
+
+  private onChangeCallback: (_: any) => void = () => { };
+  private onTouchedCallback: () => void = () => { };
+
+  constructor() { }
+
+  get value(): boolean {
+    return this.innerValue ? this.innerValue : null;
+  }
+
+  set value(v: boolean) {
+    if (v !== this.innerValue) {
+      this.innerValue = v;
+      this.indeterminate = false;
+      this.onChangeCallback(v);
+      this.onTouchedCallback();
+      this.onChange.emit(v);
+    }
+  }
+
+  writeValue(v: boolean): void {
+    if (v !== this.innerValue) {
+      this.innerValue = v;
+    }
+  }
+
+  registerOnChange(fn: any): void {
+    this.onChangeCallback = fn;
+  }
+
+  registerOnTouched(fn: any): void {
+    this.onTouchedCallback = fn;
+  }
+
+  setDisabledState?(isDisabled: boolean): void {
+    this.isDisabled = isDisabled;
+  }
+
+  _onBlur($event) {
+    this.focused = false;
+    this.onTouchedCallback();
+    this.onBlur.emit($event);
+  }
+
+  _onFocus($event) {
+    this.focused = true;
+    this.onFocus.emit($event);
+  }
+
+  focus() {
+    this.inputViewChild.nativeElement.focus();
+  }
+
+  _onChange(event: any) {
+    this.value = event.target.checked;
+  }
+
+  _onKeyDown(event) {
+    if (this.readOnly || this.isDisabled) {
+      event.preventDefault();
+    }
+  }
+
+}