+++ src/app/np-checkbox-demo/np-checkbox-demo-routing.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { NpCheckboxDemoComponent } from './np-checkbox-demo.component';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: NpCheckboxDemoComponent
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class NpCheckboxDemoRoutingModule { }